agents/pyproject.toml
[tool.poetry]
name = "doctor-auto-agents"
version = "0.1.0"
description = "Multi-agent RAG system for Doctor Auto Prime"
authors = ["Doctor Auto"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
chromadb = "^0.6.0"
openai = "^1.60.0"
anthropic = "^0.42.0"
supabase = "^2.11.0"
pydantic-settings = "^2.7.0"
httpx = "^0.28.0"
pdfplumber = "^0.11.0"
beautifulsoup4 = "^4.12.0"
apscheduler = "^3.10.0"
sentence-transformers = "^3.4.0"
python-multipart = "^0.0.18"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
